test(main): split the Team API suite by domain
`team-api-server.test.ts` was 2452 lines and the longest test file in the repo,
with one `describe` block of 1865 lines and a final block testing a different
class entirely. It is cut along the same seam as the route modules: each domain
keeps its cases beside the module that serves them, and the root file keeps only
what is true of the router as a whole - teardown, compatibility, the route
round-trip that has to see the whole shared table at once, and the status
contract the split of `#handle` had to preserve.

`host-service.test.ts` is the block that was testing `HostService`, now named
after its own subject.

A pure move: all 39 cases survive, one to one - 27 across the Team API files,
12 in `host-service.test.ts` - with no assertion added, dropped or rewritten
beyond the mechanical rename of the fixture accessor to `signIn` and the
`try/finally` teardowns that `afterEach(stopTeamApiFixtures)` now replaces.

Watched it fail twice, each against the domain file that should own it:
disabling the `sharedFiles` branch in `route-files.ts` reds
`team-api-server.files.test.ts` with "expected 404 to be 200", and disabling
the `remoteScreen.sessions` branch in `route-remote-screen.ts` reds
`team-api-server.remote-screen.test.ts` with "expected 404 to be 201".

`src/main/AGENTS.md` now says why `agent-service.*.test.ts` is the right model
here and the wrong one for `remote-server-*`: `TeamApiServer` is a class that
will not divide, every case drives real HTTP against a real listener, so the
only seam between two cases is the route they call.
